lru & reorganize code into different files
make chunk cache update lru when it's used

trying to not let single files grow too huge by separating logical
pieces into different files.
